Renaming component file to FetchTestFilesButton
